move table view to view controller and delete view
