--- solutions/devsprint-caio-santos-6/FinanceApp/Modules/ContactList/ContactListView.swift
@@ -1,85 +1,0 @@
-//
-//  ContactListView.swift
-//  FinanceApp
-//
-//  Created by Rodrigo Borges on 30/12/21.
-//
-
-import Foundation
-import UIKit
-
-class ContactListView: UIView {
-
-    static let cellSize = CGFloat(82)
-
-    private let cellIdentifier = "ContactCellIdentifier"
-
-    lazy var tableView: UITableView = {
-
-        let tableView = UITableView(frame: .zero)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.register(ContactCellView.self, forCellReuseIdentifier: self.cellIdentifier)
-        tableView.dataSource = self
-        tableView.delegate = self
-        return tableView
-    }()
-
-    init() {
-        super.init(frame: .zero)
-
-        backgroundColor = .white
-        addSubviews()
-        configureConstraints()
-
-        tableView.reloadData()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension ContactListView {
-
-    func addSubviews() {
-
-        addSubview(tableView)
-    }
-
-    func configureConstraints() {
-
-        NSLayoutConstraint.activate([
-
-            tableView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor),
-            tableView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor),
-            tableView.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor),
-        ])
-    }
-}
-
-extension ContactListView: UITableViewDataSource {
-
-    public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-
-        return 10
-    }
-
-    public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-
-        let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath) as! ContactCellView
-
-        return cell
-    }
-}
-
-extension ContactListView: UITableViewDelegate {
-
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return ContactListView.cellSize
-    }
-
-    public func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-
-    }
-}
